requirements.txt
datasets>=2.19.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
pypdf>=4.2.0
ragas>=0.1.9
tqdm>=4.66.0
jupyter>=1.0.0
ipykernel>=6.29.0
